Phase 4.5a - Editor UX polish
- Native file dialog via nfd-extended; typed-path popup kept as fallback
  when NFD_Init fails
- Right-click on a node now offers Duplicate and Delete alongside
  Per-voice
- Multi-select Delete is one undo entry: ed::BeginDelete loop wraps in
  EditHistory composite group
- Palette gains a case-insensitive filter input (matches MenuLabel and
  TypeName)
- Seeded patch routes through a Meter so master levels are visible by
  default
- New FEditHistory::BeginComposite/EndComposite; UndoStack now holds
  vector<FUndoEntry> so single edits and batch operations share the
  same shape

+++ src/ui/SequencerUI.cpp
@@ -33,7 +33,7 @@ namespace NodeSynth
 		// Lay out 16 columns. Each column has: enable checkbox, pitch
 		// vertical slider, velocity horizontal bar, gate-length bar.
 		ImGui::BeginGroup();
-		for (size_t I = 0; I < FSequencer::MaxSteps; ++I)
+		for (uint32_t I = 0; I < FSequencer::MaxSteps; ++I)
 		{
 			ImGui::PushID(static_cast<int>(I));
